feat(run/mod.rs,run/variables.rs): Made the values of each variable to be editable
